+++ application/controllers/Sales.php
@@ -75,29 +75,39 @@ class Sales extends Application
     public function sales(){
         $this->data['pagebody'] = 'sale_confirmation';
         //$inventory[] = array();
-        foreach($_POST as $post_name => $post_value){
-            $this->Transactions->setProducts($post_name, $post_value);
+        foreach($_POST as $post_name => $post_value){    
             $inventory[] = array('key' => $post_name, 'value' => $post_value);
         }
+        $outcome = array();
+        $totalSum = 0;
+        $okToProcess = 0;
+    
+        foreach($inventory as $product){
+            $record = $this->Products->get($product['key']);
+            $quantityOrdered = $product['value'];
 
-        $result = array();
-        $sum = 0;
-        foreach($inventory as $source){
-            $record = $this->Products->get($source['key']);
-
-            if($source['value'] == 1){
-                $result[] = array('line' => "You ordered " . $source['value'] . ' ' .  $record->name .
-                    " at " . $this->toDollars($record->price)  . " per unit." . "</br>");
-            }else if($source['value'] > 1){
-                $result[] = array('line' => "You ordered " . $source['value'] . ' ' .  $record->name .
-                    "s at " . $this->toDollars($record->price)  . " per unit." . "</br>");
+            if(($quantityOrdered != 0) && ($quantityOrdered <= $record->stock)){
+                $okToProcess = $quantityOrdered;   
+            }else if($quantityOrdered == 0){
+                $okToProcess = 0;
+            }else{
+                $okToProcess = -1;
             }
 
-            $sum += $record->price * $source['value'];
+            if($okToProcess == -1){
+                $outcome[] = array('line' => "Not enough stocks to process the order: " . $record->name . "</br>");
+            }else if($okToProcess == 1){
+                $outcome[] = array('line' => "You ordered " . $okToProcess . ' ' . $record->name . " at " . $this->toDollars($record->price) . "per unit." . "</br>");
+            }else if($okToProcess > 1){
+                $outcome[] = array('line' => "You ordered " . $okToProcess . ' ' . $record->name . "s at " . $this->toDollars($record->price) . "per unit." . "</br>");
+            }            
+            
+            if($okToProcess > 0){
+                $totalSum += $record->price * $okToProcess;    
+            }            
         }
-        $result[] = array('line' => "<br><strong>Grand Total:</strong> " . $this->toDollars($sum));
-        $this->data['result'] = $result;
-
+        $outcome[] = array('line' => "<br><strong>Grand Total:</strong> " . $this->toDollars($totalSum));
+        $this->data['result'] = $outcome;
         $this->render();
     }